+++ test/Sentry.Tests/Internals/Http/CachingTransportTests.cs
@@ -15,7 +15,7 @@ public class CachingTransportTests
         _logger = new TestOutputDiagnosticLogger(testOutputHelper);
     }
 
-    [Fact(Timeout = 7000)]
+    [Fact]
     public async Task WithAttachment()
     {
         // Arrange
@@ -40,7 +40,8 @@ public class CachingTransportTests
                  exception = readStreamException;
              }
          })));
-        await using var transport = CachingTransport.Create(innerTransport, options);
+
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
 
         var tempFile = Path.GetTempFileName();
 
@@ -51,7 +52,7 @@ public class CachingTransportTests
 
             // Act
             await transport.SendEnvelopeAsync(envelope);
-            await WaitForDirectoryToBecomeEmptyAsync(cacheDirectory.Path);
+            await transport.FlushAsync();
 
             // Assert
             if (exception != null)
@@ -65,7 +66,7 @@ public class CachingTransportTests
         }
     }
 
-    [Fact(Timeout = 7000)]
+    [Fact]
     public async Task WorksInBackground()
     {
         // Arrange
@@ -81,10 +82,16 @@ public class CachingTransportTests
         using var innerTransport = new FakeTransport();
         await using var transport = CachingTransport.Create(innerTransport, options);
 
+        // Attach to the EnvelopeSent event. We'll wait for that below.
+        // ReSharper disable once AccessToDisposedClosure
+        using var waiter = new Waiter<Envelope>(handler => innerTransport.EnvelopeSent += handler);
+
         // Act
         using var envelope = Envelope.FromEvent(new SentryEvent());
         await transport.SendEnvelopeAsync(envelope);
-        await WaitForDirectoryToBecomeEmptyAsync(cacheDirectory.Path);
+
+        // wait for the inner transport to signal that it sent the envelope
+        await waiter.WaitAsync(TimeSpan.FromSeconds(7));
 
         // Assert
         var sentEnvelope = innerTransport.GetSentEnvelopes().Single();
@@ -112,11 +119,11 @@ public class CachingTransportTests
 
         innerTransport
             .SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>())
-            .ThrowsForAnyArgs(_ =>
+            .ReturnsForAnyArgs(async _ =>
             {
                 capturingCompletionSource.SetResult(null);
-                cancelingCompletionSource.Task.Wait(TimeSpan.FromSeconds(4));
-                return new OperationCanceledException();
+                await Task.WhenAny(cancelingCompletionSource.Task, Task.Delay(TimeSpan.FromSeconds(4)));
+                throw new OperationCanceledException();
             });
 
         await using var transport = CachingTransport.Create(innerTransport, options);
@@ -184,8 +191,6 @@ public class CachingTransportTests
     [Fact]
     public async Task EnvelopeReachesInnerTransport()
     {
-        var timeout = TimeSpan.FromSeconds(7);
-
         // Arrange
         using var cacheDirectory = new TempDirectory();
         var options = new SentryOptions
@@ -197,25 +202,20 @@ public class CachingTransportTests
         };
 
         using var innerTransport = new FakeTransport();
-        await using var transport = CachingTransport.Create(innerTransport, options);
-
-        var tcs = new TaskCompletionSource<bool>();
-        using var cts = new CancellationTokenSource(timeout);
-        innerTransport.EnvelopeSent += (_, _) => tcs.SetResult(true);
-        cts.Token.Register(() => tcs.TrySetCanceled());
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
 
         // Act
         using var envelope = Envelope.FromEvent(new SentryEvent());
-        await transport.SendEnvelopeAsync(envelope, cts.Token);
-        var completed = await tcs.Task;
+        await transport.SendEnvelopeAsync(envelope);
+        await transport.FlushAsync();
 
         // Assert
-        Assert.True(completed, "The task timed out!");
         var sentEnvelope = innerTransport.GetSentEnvelopes().Single();
-        sentEnvelope.Should().BeEquivalentTo(envelope, o => o.Excluding(x => x.Items[0].Header));
+        sentEnvelope.Should().BeEquivalentTo(envelope,
+            o => o.Excluding(x => x.Items[0].Header));
     }
 
-    [Fact(Timeout = 5000)]
+    [Fact]
     public async Task MaintainsLimit()
     {
         // Arrange
@@ -230,28 +230,20 @@ public class CachingTransportTests
         };
 
         var innerTransport = Substitute.For<ITransport>();
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
 
-        var tcs = new TaskCompletionSource<object>();
-
-        // Block until we're done
-        innerTransport
-            .When(t => t.SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>()))
-            .Do(_ => tcs.Task.Wait());
-
-        await using var transport = CachingTransport.Create(innerTransport, options);
-
-        // Act & assert
+        // Act
         for (var i = 0; i < options.MaxCacheItems + 2; i++)
         {
             using var envelope = Envelope.FromEvent(new SentryEvent());
             await transport.SendEnvelopeAsync(envelope);
-
-            transport.GetCacheLength().Should().BeLessOrEqualTo(options.MaxCacheItems);
         }
-        tcs.SetResult(null);
+
+        // Assert
+        transport.GetCacheLength().Should().BeLessOrEqualTo(options.MaxCacheItems);
     }
 
-    [Fact(Timeout = 7000)]
+    [Fact]
     public async Task AwareOfExistingFiles()
     {
         // Arrange
@@ -265,31 +257,32 @@ public class CachingTransportTests
         };
 
         // Send some envelopes with a failing transport to make sure they all stay in cache
+        var initialInnerTransport = new FakeFailingTransport();
+        await using var initialTransport = CachingTransport.Create(initialInnerTransport, options, startWorker: false);
+
+        for (var i = 0; i < 3; i++)
         {
-            using var initialInnerTransport = new FakeTransport();
-            await using var initialTransport = CachingTransport.Create(initialInnerTransport, options);
-
-            // Shutdown the worker immediately so nothing gets processed
-            await initialTransport.StopWorkerAsync();
-
-            for (var i = 0; i < 3; i++)
-            {
-                using var envelope = Envelope.FromEvent(new SentryEvent());
-                await initialTransport.SendEnvelopeAsync(envelope);
-            }
+            using var envelope = Envelope.FromEvent(new SentryEvent());
+            await initialTransport.SendEnvelopeAsync(envelope);
         }
 
-        using var innerTransport = new FakeTransport();
-        await using var transport = CachingTransport.Create(innerTransport, options);
+        // Move them all to processing and leave them there (due to FakeFailingTransport)
+        await initialTransport.FlushAsync();
 
         // Act
-        await WaitForDirectoryToBecomeEmptyAsync(cacheDirectory.Path);
+
+        // Creating the transport should move files from processing during initialization.
+        using var innerTransport = new FakeTransport();
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
+
+        // Flushing the worker will ensure all files are processed.
+        await transport.FlushAsync();
 
         // Assert
         innerTransport.GetSentEnvelopes().Should().HaveCount(3);
     }
 
-    [Fact(Timeout = 7000)]
+    [Fact]
     public async Task NonTransientExceptionShouldLog()
     {
         // Arrange
@@ -308,10 +301,7 @@ public class CachingTransportTests
             .SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>())
             .Returns(_ => Task.FromException(new Exception("The Message")));
 
-        await using var transport = CachingTransport.Create(innerTransport, options);
-
-        // Can't really reliably test this with a worker
-        await transport.StopWorkerAsync();
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
 
         // Act
         using var envelope = Envelope.FromEvent(new SentryEvent());
@@ -328,7 +318,7 @@ public class CachingTransportTests
         Assert.Equal("Failed to send cached envelope: {0}, discarding cached envelope. Envelope contents: {1}", message);
     }
 
-    [Fact(Timeout = 7000)]
+    [Fact]
     public async Task DoesNotRetryOnNonTransientExceptions()
     {
         // Arrange
@@ -347,14 +337,12 @@ public class CachingTransportTests
         innerTransport
             .SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>())
             .Returns(_ =>
+                // ReSharper disable once AccessToModifiedClosure
                 isFailing
                     ? Task.FromException(new InvalidOperationException())
                     : Task.CompletedTask);
 
-        await using var transport = CachingTransport.Create(innerTransport, options);
-
-        // Can't really reliably test this with a worker
-        await transport.StopWorkerAsync();
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
 
         // Act
         for (var i = 0; i < 3; i++)
@@ -375,28 +363,139 @@ public class CachingTransportTests
         _ = innerTransport.Received(0).SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>());
     }
 
-    [Fact(Timeout = 7000)]
-    public async Task DoesNotDeleteCacheIfHttpRequestException()
+    [Fact]
+    public async Task RecordsDiscardedEventOnNonTransientExceptions()
     {
-        var exception = new HttpRequestException(null);
-        await TestNetworkException(exception);
+        // Arrange
+        using var cacheDirectory = new TempDirectory();
+        var options = new SentryOptions
+        {
+            Dsn = DsnSamples.ValidDsnWithoutSecret,
+            DiagnosticLogger = _logger,
+            Debug = true,
+            CacheDirectoryPath = cacheDirectory.Path,
+            ClientReportRecorder = Substitute.For<IClientReportRecorder>()
+        };
+
+
+        var innerTransport = Substitute.For<ITransport>();
+        innerTransport
+            .SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>())
+            .Returns(_ => Task.FromException(new InvalidOperationException()));
+
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
+
+        // Act
+        using var envelope = Envelope.FromEvent(new SentryEvent());
+        await transport.SendEnvelopeAsync(envelope);
+        await transport.FlushAsync();
+
+        // Test that we recorded the discarded event
+        options.ClientReportRecorder.Received(1)
+            .RecordDiscardedEvent(DiscardReason.CacheOverflow, DataCategory.Error);
     }
 
-    [Fact(Timeout = 7000)]
-    public async Task DoesNotDeleteCacheIfIOException()
+    [Fact]
+    public async Task RoundtripsClientReports()
     {
-        var exception = new IOException(null);
-        await TestNetworkException(exception);
+        // Arrange
+        using var cacheDirectory = new TempDirectory();
+        var options = new SentryOptions
+        {
+            Dsn = DsnSamples.ValidDsnWithoutSecret,
+            DiagnosticLogger = _logger,
+            Debug = true,
+            CacheDirectoryPath = cacheDirectory.Path
+        };
+
+        var timestamp = DateTimeOffset.UtcNow;
+        var clock = Substitute.For<ISystemClock>();
+        clock.GetUtcNow().Returns(timestamp);
+        var recorder = new ClientReportRecorder(options, clock);
+        options.ClientReportRecorder = recorder;
+
+        var innerTransport = new FakeTransportWithRecorder(recorder);
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
+
+        // Act
+        recorder.RecordDiscardedEvent(DiscardReason.BeforeSend, DataCategory.Error);
+        using var envelope = Envelope.FromEvent(new SentryEvent());
+        await transport.SendEnvelopeAsync(envelope); // should internally generate a client report and write to disk
+        var interimClientReport = recorder.GenerateClientReport(); // should be null
+        await transport.FlushAsync(); // will read from disk and should send that client report
+
+        // Test that the interim report was null
+        Assert.Null(interimClientReport);
+
+        // Test that we actually sent the discarded event
+        var clientReport = new ClientReport(timestamp,
+            new Dictionary<DiscardReasonWithCategory, int>
+            {
+                {DiscardReason.BeforeSend.WithCategory(DataCategory.Error), 1}
+            }
+        );
+        var expected = await EnvelopeItem.FromClientReport(clientReport).Payload.SerializeToStringAsync(_logger);
+
+        var envelopeSent = innerTransport.GetSentEnvelopes().Single();
+        var envelopeItem = envelopeSent.Items.Single(x => x.TryGetType() == "client_report");
+        var actual = await envelopeItem.Payload.SerializeToStringAsync(_logger);
+        Assert.Equal(expected, actual);
     }
 
-    [Fact(Timeout = 7000)]
-    public async Task DoesNotDeleteCacheIfSocketException()
+    [Fact]
+    public async Task RestoresDiscardedEventCounts()
     {
-        var exception = new SocketException();
-        await TestNetworkException(exception);
+        // Arrange
+        using var cacheDirectory = new TempDirectory();
+        var options = new SentryOptions
+        {
+            Dsn = DsnSamples.ValidDsnWithoutSecret,
+            DiagnosticLogger = _logger,
+            Debug = true,
+            CacheDirectoryPath = cacheDirectory.Path
+        };
+
+        var recorder = (ClientReportRecorder) options.ClientReportRecorder;
+        var innerTransport = new FakeTransportWithRecorder(recorder);
+        await using var transport = CachingTransport.Create(innerTransport, options,
+            startWorker: false, failStorage: true);
+
+        // some arbitrary discarded events ahead of time
+        recorder.RecordDiscardedEvent(DiscardReason.BeforeSend, DataCategory.Attachment);
+        recorder.RecordDiscardedEvent(DiscardReason.EventProcessor, DataCategory.Error);
+        recorder.RecordDiscardedEvent(DiscardReason.EventProcessor, DataCategory.Error);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+        recorder.RecordDiscardedEvent(DiscardReason.QueueOverflow, DataCategory.Security);
+
+        // Act
+        await Assert.ThrowsAnyAsync<Exception>(async () =>
+        {
+            using var envelope = Envelope.FromEvent(new SentryEvent());
+            await transport.SendEnvelopeAsync(envelope); // will fail, since we set failStorage to true
+        });
+
+        // Assert
+        recorder.DiscardedEvents.Should().BeEquivalentTo(new Dictionary<DiscardReasonWithCategory, int>
+        {
+            // These are the original items recorded.  They should still be there.
+            {DiscardReason.BeforeSend.WithCategory(DataCategory.Attachment), 1},
+            {DiscardReason.EventProcessor.WithCategory(DataCategory.Error), 2},
+            {DiscardReason.QueueOverflow.WithCategory(DataCategory.Security), 3}
+        });
     }
 
-    private async Task TestNetworkException(Exception exception)
+    public static IEnumerable<object[]> NetworkTestData =>
+        new List<object[]>
+        {
+            new object[] {new HttpRequestException(null)},
+            new object[] {new IOException(null)},
+            new object[] {new SocketException()}
+        };
+
+    [Theory]
+    [MemberData(nameof(NetworkTestData))]
+    public async Task TestNetworkException(Exception exception)
     {
         // Arrange
         using var cacheDirectory = new TempDirectory();
@@ -415,10 +514,7 @@ public class CachingTransportTests
             .SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>())
             .Returns(_ => Task.FromException(exception));
 
-        await using var transport = CachingTransport.Create(innerTransport, options);
-
-        // Can't really reliably test this with a worker
-        await transport.StopWorkerAsync();
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
 
         using var envelope = Envelope.FromEvent(new SentryEvent());
         await transport.SendEnvelopeAsync(envelope);
@@ -442,41 +538,5 @@ public class CachingTransportTests
         // Assert
         Assert.Equal(exception, receivedException);
         Assert.True(Directory.EnumerateFiles(cacheDirectory.Path, "*", SearchOption.AllDirectories).Any());
-    }
-
-    private async Task WaitForDirectoryToBecomeEmptyAsync(string directoryPath, TimeSpan? timeout = null)
-    {
-        bool DirectoryIsEmpty() => !Directory.EnumerateFiles(directoryPath, "*", SearchOption.AllDirectories).Any();
-
-        if (DirectoryIsEmpty())
-        {
-            // No point in waiting if the directory is already empty
-            return;
-        }
-
-        using var cts = new CancellationTokenSource(timeout ?? TimeSpan.FromSeconds(7));
-
-        using var watcher = new FileSystemWatcher(directoryPath);
-        watcher.IncludeSubdirectories = true;
-        watcher.EnableRaisingEvents = true;
-
-        // Wait until timeout or directory is empty
-        while (!DirectoryIsEmpty())
-        {
-            cts.Token.ThrowIfCancellationRequested();
-
-            var tcs = new TaskCompletionSource<bool>();
-            cts.Token.Register(() => tcs.TrySetCanceled());
-            watcher.Deleted += (_, _) => tcs.TrySetResult(true);
-
-            // One final check before waiting
-            if (DirectoryIsEmpty())
-            {
-                return;
-            }
-
-            // Wait for a file to be deleted, but not longer than 100ms
-            await Task.WhenAny(tcs.Task, Task.Delay(100, cts.Token));
-        }
     }
 }